Retrieve BO item proerties

Get the properties of items as tibble.

Usage

get_bo_item(
  conn,
  name = NULL,
  parent_folder = NULL,
  kind = NULL,
  owner = NULL,
  id = NULL
)

Arguments

conn

Connection reference

name

Name of the item (optional). You can use SQL wilcard % in the name.

parent_folder

Numeric id of the parent folder (optional)

kind

Kind of item (optional)

owner

Owner of item (optional)

id

Id of item (optional)

Value

Tibble of item properties

Details

After opening a connection with open_bo_connection you can retrieve the properties of a items with this function. The properties returned contain the item id in the SI_ID column. This matches the ID in the properties page in the BO web GUI. If you have the id of the parent folder that will narrow the search in case there are duplicate item names. If more than one item is returned by the query the tibble will contain multiple rows

Examples

# get the id of an item. Normally you would call open_bo_connection()
conn= open_bo_connection(); get_bo_item(conn, name='test',parent_folder=3944223)$SI_ID
#> [1] 9104971
# get all the Excel documents in a folder
get_bo_item(open_bo_connection(),parent_folder = 3944223, kind = 'Excel')
#> # A tibble: 10 × 7
#>      SI_ID SI_NAME                       SI_UP…¹ SI_CR…² SI_KIND SI_OW…³ SI_PA…⁴
#>      <int> <chr>                         <chr>   <chr>   <chr>   <chr>     <int>
#>  1 5890750 BoExcelVisibilityTest         May 7,… May 7,… Excel   john.c… 3944223
#>  2 5893595 Retail Promo Snapshot PG 156… Oct 31… May 7,… Excel   john.c… 3944223
#>  3 9057978 Retail Promo Snapshot PG 156… Oct 28… Oct 28… Excel   bowebr… 3944223
#>  4 9086503 test52cop%y.xlsx              Nov 7,… Nov 2,… Excel   john.c… 3944223
#>  5 9086772 test52.xlsx                   Nov 5,… Nov 2,… Excel   bowebr… 3944223
#>  6 9104971 test                          Nov 7,… Nov 5,… Excel   bowebr… 3944223
#>  7 9122562 test.xlsx                     Nov 20… Nov 8,… Excel   bowebr… 3944223
#>  8 9154733 test2.xlsx                    Nov 14… Nov 14… Excel   bowebr… 3944223
#>  9 9162554 zQ22022/Q2 Customer Joined D… Nov 15… Nov 15… Excel   bowebr… 3944223
#> 10 9189933 text2.xlsx                    Nov 20… Nov 20… Excel   bowebr… 3944223
#> # … with abbreviated variable names ¹​SI_UPDATE_TS, ²​SI_CREATION_TIME,
#> #   ³​SI_OWNER, ⁴​SI_PARENT_FOLDER