Open a connection to a BO server. You can keep parameters in environment variables

Open a connection to a BO server. You can keep parameters in environment variables

Usage

open_bo_connection(
  server = Sys.getenv("BO_SERVER"),
  username = Sys.getenv("BO_USERNAME"),
  password = get_user_password(username, server),
  save_password = TRUE
)

Arguments

server

Server to connect to "server:port" (optional) Defaults BO_SERVER environment variable

username

(optional) Defaults to "BO_USERNAME environment variable

password

(optional)

save_password

Save the password encrypted in the database. Saving password is required for automatic reconnection on token expiration

Value

Connection reference

Examples

# open a connection to the server in environment variable "BO_SERVER" using user in "BO_USERNAME".
# and password in the database
conn <- open_bo_connection()
#> Error: Could not resolve host: biprws
# open a connection to a server using username and password
conn <- open_bo_connection(Sys.getenv("BO_SERVER"), username = 'john.capehart', password=getPass())
#> Error in getPass(): could not find function "getPass"